`l1_handler_execute` cheatcode (#459)
<!-- Reference any GitHub issues resolved by this PR -->

Closes #334 (and #347 ?)

## Introduced changes

This PR adds a cheatcode to reproduce what the sequencer does when
a message is received from ethereum.

For now, I have used the selector of the `#[l1_handler]` instead
of a cairo short string as in solidity code we use the selector, it
looks
more natural to me.
If you prefer cairo short string from a UX perspective, I'll switch it.
:+1:

## Breaking changes

New cheatcode.

## Checklist

<!-- Make sure all of these are complete -->

- [X] Linked relevant issue
- [X] Updated relevant documentation
- [X] Added relevant tests
- [X] Performed self-review of the code
- [X] Added changes to `CHANGELOG.md`