Fix wrong error message when there is no scarb and path to scarb.toml is supplied (#607)
<!-- Reference any GitHub issues resolved by this PR -->

Closes https://github.com/foundry-rs/starknet-foundry/issues/545

## Introduced changes

When there was no scarb in PATH and a user passed the
`--path-to-scarb-toml`, the error was that we couldn't find the file
(even though it exists in the fs), where is should be that scarb was not
found in PATH. This PR changes that behaviour. The issue did not
persists, if user did not pass the `--path-to-scarb-toml` flag.

## Breaking changes

None

## Checklist

<!-- Make sure all of these are complete -->

- [X] Linked relevant issue
- [X] Updated relevant documentation
- [X] Added relevant tests
- [X] Performed self-review of the code
- [X] Added changes to `CHANGELOG.md`