[tests] Use dynamically calculated hashes and addresses instead of hard-coded ones (#650)
<!-- Reference any GitHub issues resolved by this PR -->

This PR does not close any issue, but is a part of
https://github.com/foundry-rs/starknet-foundry/issues/429
Hashes and addresses changed every time we updated devnet/compilers.
This is a bit tiresome to update them each time, plus there probably is
no need for that. This PR makes the tests to use hashes and addresses
taken directly from our code.

## Introduced changes

<!-- A brief description of the changes -->

- do not use hard coded hashes and addresses

## Breaking changes

<!-- List of all breaking changes, if applicable -->
None

## Checklist

<!-- Make sure all of these are complete -->

- [X] Linked relevant issue
- [X] Updated relevant documentation
- [X] Added relevant tests
- [X] Performed self-review of the code
- [X] Added changes to `CHANGELOG.md`