Fix live printing (#880)
<!-- Reference any GitHub issues resolved by this PR -->

Closes #860
Closes #843

## Introduced changes

<!-- A brief description of the changes -->

- Test from different crates no longer run in parallel
- Tests from crate are now run unordered.
- Outputs from these tests are printed immediately after test case
execution completed
- Added new macro for testing e2e tests outputs that does not depend on
test order. This macro is limited right now in that it asserts the
output line by line. This is problematic for outputs that are longer
than one line, like the whole output of `[FAIL]` test. This will be
extended in future PRs.

## Breaking changes

<!-- List of all breaking changes, if applicable -->

## Checklist

<!-- Make sure all of these are complete -->

- [x] Linked relevant issue
- [x] Updated relevant documentation
- [x] Added relevant tests
- [x] Performed self-review of the code
- [x] Added changes to `CHANGELOG.md`

(cherry picked from commit 3aba230c6107379891218749ebdecf73db54cc57)