Simplify cancelation logic in the context of parallel execution (#989)
<!-- Reference any GitHub issues resolved by this PR -->
The main aid of this PR is to simplify cancelation logic. The previous
version was base on the cancellation token. In order to optimize cases
when tests could be skipped. Because tokio::select doesn't work with
`spawn_blocking` logic with channels was added, to gently close the
blocking task, after cancelled tokio::tasks. After deeper research, we
realized that communication-based on channels and "closing" blocking
spawn task is enough.

Currently, if we use `--exit-first` flag, the failing test will close
the channel, so all next spawn_blocking tasks will return `Skipped`
instead of executing test

## Introduced changes

<!-- A brief description of the changes -->

-
-

## Breaking changes

<!-- List of all breaking changes, if applicable -->

## Checklist

<!-- Make sure all of these are complete -->

- [x] Linked relevant issue
- [x] Updated relevant documentation
- [x] Added relevant tests
- [x] Performed self-review of the code
- [x] Added changes to `CHANGELOG.md`