feat: assert_not_emitted (#1425)
<!-- Reference any GitHub issues resolved by this PR -->

Closes #1409

## Introduced changes

Adding a `assert_not_emitted` check to verify an event was *not*
emitted. The counterparty to `assert_emitted`.

I've also renamed the `assert_if_emitted` helper function to
`is_emitted` because 1) it didn't assert and 2) there's no operation on
"if emitted". I'm guessing it's just a relic from some refactoring.

## Checklist

<!-- Make sure all of these are complete -->

- [x] Linked relevant issue
- [x] Updated relevant documentation
- [x] Added relevant tests
- [x] Performed self-review of the code
- [x] Added changes to `CHANGELOG.md`