Add incompatible RPC version warning to sncast (#1704)
<!-- Reference any GitHub issues resolved by this PR -->

Closes #1560 

## Introduced changes

<!-- A brief description of the changes -->

- Print a warning message about the incompatible RPC version when
executing `sncast` commands

Note: The logic is partially tested on the forge side. For comprehensive
testing in sncast, we would need to either use one of our nodes or use
mocks. We may also consider moving more code to the shared crate, please
share your thoughts on any improvements in this regard.

## Checklist

<!-- Make sure all of these are complete -->

- [X] Linked relevant issue
- [ ] Updated relevant documentation
- [ ] Added relevant tests
- [X] Performed self-review of the code
- [X] Added changes to `CHANGELOG.md`

---------

Co-authored-by: Piotr Figiela <77412592+Draggu@users.noreply.github.com>