Save sierra test code separately for profiler (#2065)
## Introduced changes

- Test artifacts file usually had 2 structs in it:
`CompiledTestCrateRaw` for `src/` and `CompiledTestCrateRaw` for
`tests/`, each with a separate sierra program. This made profiler unable
to determine which sierra it should use for function level profiling
(only path to the artifact file was passed to profiler). Now we save the
relevant sierra program to another file and pass a path to the trace.
This way profiler knows what sierra it should use. It also makes
profiler less coupled with snforge (it doesn't have to know forge test
artifact format)

## Checklist

- [x] Linked relevant issue
- [x] Updated relevant documentation
- [x] Added relevant tests
- [x] Performed self-review of the code
- [x] Added changes to `CHANGELOG.md`