Contract verification (#2252)
Closes #892 

## Introduced changes

This PR builds upon the changes introduced in PR #875. The original PR
is still open because the Voyager APIs are not ready yet, but we require
the verification feature for the Walnut Debugger. To accommodate this,
we have incorporated the changes from PR #875 and made the following
additions:
* Introduced the `WalnutVerificationInterface`.
* Resolved branch conflicts.
* Made minor modifications to ensure compatibility.

Furthermore, to make the PR merge-ready, we have temporarily removed the
Voyager and Starkscan interfaces since their respective APIs are not
currently available. New verification interfaces can be added once
Voyager, Starkscan, or any other provider adds API support for Cairo
contract verification.

This Pull Request adds the verify command for the contract verification.
More about the details of the contract verification can found
[here](https://github.com/foundry-rs/starknet-foundry/blob/master/design_documents/contract_verification.md).

* Added `verify` command to the `sncast` tool
* Modified `cast/main.rs` & created
`cast/src/starknet_commands/verify.rs` mainly

## Checklist

- [x] Linked relevant issue
- [x] Updated relevant documentation
- [x] Added relevant tests
- [x] Performed self-review of the code
- [x] Added changes to `CHANGELOG.md`

---------

Co-authored-by: MBarwicki <mbarwicki@bravurasolutions.com>
Co-authored-by: Rohit Ranjan <rohitrjn629@gmail.com>
Co-authored-by: Tomasz Rejowski <34059895+Arcticae@users.noreply.github.com>
Co-authored-by: Karol Sewiło <95349104+ksew1@users.noreply.github.com>