Fix data transformer integration (#2624)
<!-- Reference any GitHub issues resolved by this PR -->

Closes #2588

## Introduced changes

<!-- A brief description of the changes -->

Previous data transformer integration used `--calldata` argument (along
with the old serialized input) and used `()` to indicate data used for
transformation. `()` are special characters in bash so it did not work
in most shells, requiring wrapping in `''` anyway. This PR changes this
integration so data for transformation is passed as a single string to a
new `--arguments` argument and doesn't require any special wrapping.

- Changed the syntax of data transformer so it works better with shells
- Introduced new `--arguments` argument to `sncast` to distinguish data
transformation
- Fixed support for shortstrings in data transformer
- Added tests using `.sh` scripts for data transformation

## Checklist

<!-- Make sure all of these are complete -->

- [x] Linked relevant issue
- [x] Updated relevant documentation
- [x] Added relevant tests
- [x] Performed self-review of the code
- [x] Added changes to `CHANGELOG.md`