Add rpc argument to verify subcommand
Unfortunately I wasn't able to just use flatten out `RpcArgs` in the
`Verify` structure due to the clashing `network` argument.  It has
slightly different meaning in the `RpcArgs` than in the `verify`
subcommand.  The rpc one targets only rpc server while the one in the
`verify` helps to select appropriate block explorer and must be
present.

As a side note is the logic for coming up with `JsonRpcClient` in the
`RpcArgs` correct?  I would assume that if I pass a manual `--url`
argument that would overwrite everything, but it seems in the
`RpcArgs` the `--network` argument takes precedence.