Pin rust version to 1.86.0 due to `cairo_vm` incompatibility (#3352)
<!-- Reference any GitHub issues resolved by this PR -->

Closes #

## Introduced changes

<!-- A brief description of the changes -->

Interface of `is_multiple_of` changed in rust 1.87.0. This is a know
regression that won't be resolved, see
https://github.com/rust-lang/rust/issues/140020

Due to that, `cairo_vm` no longer compiles on later rust version. Until
they upgrade and we upgrade our version of `cairo_vm`, then we have to
stay at rust 1.86.0

Issue to remove the pinned version
https://github.com/foundry-rs/starknet-foundry/issues/3351

## Checklist

<!-- Make sure all of these are complete -->

- [x] Linked relevant issue
- [x] Updated relevant documentation
- [x] Added relevant tests
- [x] Performed self-review of the code
- [x] Added changes to `CHANGELOG.md`